Create public-facing navigation bar (#1)
* Create public-facing navigation bar

Includes a logo with my name, and links to portfolio and blog pages. Also rearranged some of the template layouts, creating a public folder for public nav and template and a private folder for the admin panel which will feature options to modify portfolio/blog items.

* Fix profile picture component

The profile picture component was not reading relative to root before, which would mean it would not be found on non-root pages.

* Fix failing test and profile settings

Also renamed management to admin-page. Fix added livewire config and set default component page to this. I may change this in the future as it feels a little counterintuitve to have admin as default and I'm not sure I like this kind of inexplicit default.