Implement admin for access control (#6)
* Add migration for `is_admin` in users table

This is implemented as a simple boolean which defaults to `false`.

* Add IsAdminMiddleware

Allows requests if the user is authorised and has the `is_admin` field set to `true`, otherwise returns HTTP response code 403 for 'forbidden'.

* Restrict management routes to admins

* Add PostPolicy use in post components

Post policy allows views by any guest or user, with CRUD operations being admin-only. This is used in the Livewire component actions to add an additional layer of security.

* Update tests for access control

Adapt post and management tests for access control. This included updating old tests to use admin where necessary, and adding new tests to ensure a non-admin user cannot perform admin-only requests.

Additional changes:

 - Reduced the size of randomly created posts in `PostFactory`
 - Added an `admin()` method to `UserFactory`, making it easier to create a test admin user
 - Renamed DashboardTest ManagementTest to line up with the changed URL
 - Increased memory limit for tests, since they started to run out