Implement blog post attachment management (#8)
* Add component for attachment management and writers

`AttachmentWriter` is an interface providing attachment writing facilities. I have made two implementations:

1. `BufferedAttachmentWriter`: acts as a buffer, storing any files uploaded to it in memory until `upload()` is called at which point it is written to disk
2. `DirectAttachmentWriter`: writes any file uploaded directly to disk

The respective use cases for these at the moment is:

1. `BufferedAttachmentWriter`: used in post creation, since we do not yet have an area to write them to until the user clicks 'add'
2. `DirectAttachmentWriter`: used in post edit page, avoiding complexity of buffering edit/delete operations etc

I also add the attachment manager, a Livewire component which provides a user interface for adding and managing attachments. At the moment, users can only add attachments on this interface. Edit and delete are to come next. These are currently in use on the post edit and create pages using their relevant attachment writer implementations.

* Add back try catch on store method

This was experimentally deleted and I forgot to undo it in the last commit.

* Extend attachment manager with edit and delete

Made attachment cell component, which interacts with its parent attachment manager via events to edit or delete the attachment.

* Create copy link button for attachments

Copies the link for the attachment, relative to the public folder. This is so that it can be used in the content.

* Update post tests for new storage path

`blog:{id}.md` -> `blog:{id}/content.md`

* Add tests for buffered attachment writer

Added tests for each method in buffered attachment writer, and fixed some bugs found as a result of it.

Also added a reflectProperty helper method in the test case base class.

* Add tests for direct attachment writer

Added tests for each method in direct attachment writer, and fixed some bugs found as a result of it. Also switched it to use a set-like structure, which is actually a map of '{name} => true' since PHP does not have a specific set structure.

* Add feature tests for attachment cell/manager

Added feature tests for attachment cell, attachment manager, and post create/edit. Also added helper methods for actingAsUser/Admin.

* Add a few more feature tests for manager

Test errors, and made a change to add to use field errors instead of session flash.