Add management pages for portfolio
Add pages to create/update projects in the portfolio, as well as an index management page displaying a list of projects with links to the create/update pages and a delete button.

Add cover image handling for creation and updating including renaming and editing the file itself. Added a 'portfolio' disk for the cover images to be stored in. Also updated migration for projects to store the filename rather than path for the cover image, since this is all we need.