Implement portfolio backend and management (#10)
* Add database migration for projects

* Add eloquent model for project

* Add slug to project migration

* Add management pages for portfolio

Add pages to create/update projects in the portfolio, as well as an index management page displaying a list of projects with links to the create/update pages and a delete button.

Add cover image handling for creation and updating including renaming and editing the file itself. Added a 'portfolio' disk for the cover images to be stored in. Also updated migration for projects to store the filename rather than path for the cover image, since this is all we need.

* Add project factory and cell test

Also added missing fake storage for post cell test.

* Add project index tests

* Create project component tests

Also fixed formatting of the component's $rules property to be consistent with the others.

* Add blog and portfolio data to gitignore

* Add edit project test

Also made fixes as a result of developing the tests, ran composer update, made cover image filename unique and made some minor aesthetic adjustments to the code.