Fix dark mode on homepage
Quick fix to set text colours properly for dark mode on the home page. Also made the GitHub icon adapt to the theme, which I have had to do with `dark:` Tailwind breakpoint for now as I cannot find a way for the `<picture>` tag to respect the appearance selected in preferences if it differs from the system one.