Convert md to html server-side
A library called Parsedown is used to do this and was added via composer. The html file is written alongside the markdown and updated alongside the markdown.

Also added admin panel on show post page, which allows you to delete a post directly or go straight to the edit page. This uses a new @isAdmin directive which only shows it if the user is logged in and an admin.