Restrict registration using tokens
Registration now requires a token which is generated by admins and sent directly to the new user via email as a link. The user is required to use this email to register and may be given the admin role if assigned.

Currently, tokens are displayed to the admin on the user page in which they are generated since emails have not been implemented yet.