Implement registration management (#21)
* Add user management page

Allows admins to view users, modify whether they are admin, and delete them.

Also added form for adding users, but this is not completed yet.

* Restrict registration using tokens

Registration now requires a token which is generated by admins and sent directly to the new user via email as a link. The user is required to use this email to register and may be given the admin role if assigned.

Currently, tokens are displayed to the admin on the user page in which they are generated since emails have not been implemented yet.

* Add registration link email

Implement the email for registration, and add a preview page for it.

* Unit tests for token repository

* Add user index and cell tests

* Update registration test

Fix small bug with token error.

* Remove email verification

Since registration requires confirming a registration code we no longer need the standard email verification, since every new user is going to be considered verified by default.