Allow projects to be marked closed source
Added checkbox to add/edit project pages for whether the project is open source, which shows/hides the repository link. It will also be marked closed source if no repository link is given. The repository link has been made optional and is set to null when the project is closed source.

On the portfolio page, 'Closed source' is displayed where the GitHub link usually appears for open source projects.